requirements-dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements-dev.in --output-file requirements-dev.txt
argcomplete==3.5.2
    # via nox
cfgv==3.4.0
    # via pre-commit
colorlog==6.9.0
    # via nox
coverage==7.6.9
    # via pytest-cov
distlib==0.3.9
    # via virtualenv
filelock==3.16.1
    # via virtualenv
identify==2.6.3
    # via pre-commit
iniconfig==2.0.0
    # via pytest
nodeenv==1.9.1
    # via pre-commit
nox==2024.10.9
    # via -r requirements-dev.in
numpy==2.2.0
    # via
    #   -r requirements.in
    #   numpy-stl
numpy-stl==3.2.0
    # via -r requirements.in
packaging==24.2
    # via
    #   nox
    #   pytest
platformdirs==4.3.6
    # via virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==4.0.1
    # via -r requirements-dev.in
pytest==8.3.4
    # via
    #   -r requirements-dev.in
    #   pytest-cov
    #   pytest-mock
    #   pytest-randomly
pytest-cov==6.0.0
    # via -r requirements-dev.in
pytest-mock==3.14.0
    # via -r requirements-dev.in
pytest-randomly==3.16.0
    # via -r requirements-dev.in
python-utils==3.9.1
    # via numpy-stl
pyyaml==6.0.2
    # via pre-commit
ruff==0.8.2
    # via -r requirements-dev.in
tomli==2.2.1
    # via coverage
typing-extensions==4.12.2
    # via python-utils
virtualenv==20.28.0
    # via
    #   nox
    #   pre-commit
